requirements.txt
gymnasium==1.0.0
gymnasium_robotics==1.3.1
h5py==3.12.1
hydra-core==1.3.2
mani_skill2==0.5.3
matplotlib==3.10.3
numpy==2.2.6
omegaconf==2.3.0
opencv_python==4.10.0.84
pygame==2.6.1
scipy==1.15.3
stable_baselines3==2.4.1
torch==2.5.1
tqdm==4.67.1
